DTO update, user stats

+++ backend/src/main/java/com/webtech/quicksketch/dto/response/ReactResponse.java
@@ -4,6 +4,6 @@ import com.webtech.quicksketch.model.enums.ReactionType;
 
 public record ReactResponse(
     ReactionType currentReaction,
-    Long totalLikes,
-    Long totalDislikes
+    int totalLikes,
+    int totalDislikes
 ) {}
